+++ tests/e2e/signalling/mod.rs
@@ -245,8 +245,8 @@ impl StreamHandler<Result<Frame, WsProtocolError>> for TestMember {
                         | Event::TracksUpdated { peer_id, .. } => {
                             assert!(self.known_peers.contains(peer_id))
                         }
-                        Event::PeersRemoved { .. } => {}
-                        Event::SnapshotSynchronized { .. } => {}
+                        Event::PeersRemoved { .. }
+                        | Event::SnapshotSynchronized { .. } => {}
                     }
                     let mut events: Vec<&Event> = self.events.iter().collect();
                     events.push(&event);
